+++ src/lib/jsonSchema/defaultValues.ts
@@ -71,8 +71,8 @@ function _defaultValues(schema: JSONSchema, isOptional: boolean, path: string[])
 	if (info.types.length > 1) {
 		throw new SchemaError('Default values cannot have more than one type.', path);
 	} else if (info.types.length == 0) {
-		console.warn('No type or format for property:', path);
-		//console.dir(schema, { depth: 10 });
+		//console.warn('No type or format for property:', path); //debug
+		//console.dir(schema, { depth: 10 }); //debug
 		return undefined;
 	}