+++ CHANGELOG.md
@@ -28,3 +28,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `superForm.isTainted(path?)`
+- `superValidate.options.allowFiles`